Automatically backed up by Learn

+++ lib/bachelor.rb
@@ -52,7 +52,7 @@ def get_average_age_for_season(data, season)
   count = 0
   data.each { |season, contestants| 
     contestants.each { |contestant| 
-      avg += (contestant["age"]).to_i
+      avg += (contestant["age"]).to_f
       count += 1
     }
   }
